requirements.txt
langchain>=0.1.14
langgraph>=0.0.30
streamlit>=1.32.0
chromadb>=0.4.24
faiss-cpu>=1.7.4
pypdf>=3.17.4
groq>=0.2.2
tavily-python>=0.3.0
serper>=0.2.0
google-generativeai>=0.3.2
python-dotenv>=1.0.1
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
sentence-transformers>=2.6.1
Flask
flask-cors 